requirements.txt
cloudmesh-common
eel
humanize
streamlit
watchdog